pyproject.toml
[project]
name = "startup-success-predictor"
version = "0.1.0"
description = "GAN-Augmented Startup Success Predictor using WGAN-GP"
authors = [
    {name = "Хузин Эльдар Русланович", email = "khuzin.er@phystech.edu"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.1.0",
    "lightning>=2.1.0",
    "hydra-core>=1.3.0",
    "pydantic-settings>=2.0.0",
    "polars>=0.20.0",
    "pyarrow>=14.0.0",
    "mlflow>=2.9.0",
    "dvc>=3.0.0",
    "kaggle>=1.5.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "onnxscript>=0.1.0",
    "torchmetrics>=1.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PTH", # flake8-use-pathlib
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["startup_success_predictor"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"

[[tool.mypy.overrides]]
module = [
    "lightning.*",
    "hydra.*",
    "mlflow.*",
    "dvc.*",
    "kaggle.*",
    "onnx.*",
    "onnxruntime.*",
]
ignore_missing_imports = true
